refactor: ui별 adapter 패키지 생성해 해당 클래스 이동
